+++ apps/modernization-ui/src/apps/patient/add/extended/inputs/address/AddressView.spec.tsx
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/react';
-import { AddressEntry } from 'apps/patient/data/entry';
+import { AddressEntry } from 'apps/patient/data';
 import { AddressView } from './AddressView';
 
 const entry: AddressEntry = {
