+++ apps/modernization-ui/src/apps/patient/data/address/index.ts
@@ -1,0 +1,5 @@
+export { initial } from './entry';
+export type { AddressEntry } from './entry';
+
+export { AddressEntryFields } from './AddressEntryFields';
+export { asAddress } from './asAddress';
